# cotacao-crypto-api/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CriptomoedaDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"codigo": {
						"type": "string"
					},
					"nome": {
						"type": "string"
					},
					"descricao": {
						"type": "string"
					},
					"cotacao_compra": {
						"type": "number",
						"format": "double"
					},
					"cotacao_venda": {
						"type": "number",
						"format": "double"
					},
					"variacao": {
						"type": "number",
						"format": "double"
					},
					"timestamp_atualizacao": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"codigo",
					"nome",
					"descricao",
					"cotacao_compra",
					"cotacao_venda",
					"variacao",
					"timestamp_atualizacao"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCriptomoedaDTO": {
				"properties": {
					"codigo": {
						"type": "string"
					},
					"nome": {
						"type": "string"
					},
					"descricao": {
						"type": "string"
					},
					"cotacao_compra": {
						"type": "number",
						"format": "double"
					},
					"cotacao_venda": {
						"type": "number",
						"format": "double"
					},
					"variacao": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"codigo",
					"nome",
					"descricao",
					"cotacao_compra",
					"cotacao_venda",
					"variacao"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCriptomoedaDTO": {
				"properties": {
					"nome": {
						"type": "string"
					},
					"descricao": {
						"type": "string"
					},
					"cotacao_compra": {
						"type": "number",
						"format": "double"
					},
					"cotacao_venda": {
						"type": "number",
						"format": "double"
					},
					"variacao": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"nome",
					"descricao",
					"cotacao_compra",
					"cotacao_venda",
					"variacao"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCotacaoCriptomoedaDTO": {
				"properties": {
					"cotacao_compra": {
						"type": "number",
						"format": "double"
					},
					"cotacao_venda": {
						"type": "number",
						"format": "double"
					},
					"variacao": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"cotacao_compra",
					"cotacao_venda",
					"variacao"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "cotacao-crypto-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/criptomoedas": {
			"get": {
				"operationId": "Show",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CriptomoedaDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CriptomoedaDTO"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCriptomoedaDTO"
							}
						}
					}
				}
			}
		},
		"/criptomoedas/{codigo}": {
			"get": {
				"operationId": "Find",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CriptomoedaDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "codigo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Destroy",
				"responses": {
					"204": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "codigo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CriptomoedaDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "codigo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCriptomoedaDTO"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateCotacao",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CriptomoedaDTO"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "codigo",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCotacaoCriptomoedaDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}